Revert (most of) "Adding in reference for collaboration"
This reverts (the silly parts of) commit e9038e7fdbed236562c01a35f080e7d229cab392.

It seems the intention was to add an item to the git ignore file. What
happened in the UFO directory got nuked and a Glyphs format file
masquerading around with a .ufo extension got put in it's place.

Ya I don't know either.